android/src/main/kotlin/com/onegini/mobile/sdk/flutter/useCases/ChangePinUseCase.kt
package com.onegini.mobile.sdk.flutter.useCases

import com.onegini.mobile.sdk.android.client.OneginiClient
import com.onegini.mobile.sdk.android.handlers.OneginiChangePinHandler
import com.onegini.mobile.sdk.android.handlers.error.OneginiChangePinError
import io.flutter.plugin.common.MethodChannel

class ChangePinUseCase(private var oneginiClient: OneginiClient) {
    operator fun invoke(result: MethodChannel.Result) {
        oneginiClient.userClient.changePin(object : OneginiChangePinHandler {
            override fun onSuccess() {
                result.success(true)
            }

            override fun onError(error: OneginiChangePinError) {
                result.error(error.errorType.toString(), error.message, null)
            }
        })
    }
}